backend/requirements.txt
# Cygnusa Guardian - Backend Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3

# PDF Processing
pdfplumber==0.10.3

# AI APIs
google-generativeai==0.3.2
openai==1.10.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
# Required for Linux environments (Railway)
python-magic==0.4.27
python-docx==1.1.0
supabase>=2.10.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx==0.27.2
